+++ dedupe/clustering.py
@@ -183,8 +183,9 @@ def greedyMatching(dupes, threshold=0.5):
     A = set()
     B = set()
 
-    dupes = ((pair, score) for pair, score in dupes if score >= threshold)
-    dupes = sorted(dupes, key=lambda score: score[1], reverse=True)
+    dupes = dupes[dupes['score'] >= threshold]
+    dupes.sort(order='score')
+    dupes = dupes[::-1]
 
     for (a, b), score in dupes:
         if a not in A and b not in B:
@@ -194,15 +195,13 @@ def greedyMatching(dupes, threshold=0.5):
             yield (a, b), score
 
 
-def gazetteMatching(dupes, threshold=0.5, n_matches=1):
-    messy_id = lambda match: match[0][0]
-    score = lambda match: match[1]
-    
-    dupes = ((pair, score) for pair, score in dupes if score >= threshold)
-    dupes = sorted(dupes, key=lambda match: (messy_id(match), -score(match)))
+def gazetteMatching(scored_blocks, n_matches=1):
 
-    for _, matches in itertools.groupby(dupes, key=messy_id):
+    for block in scored_blocks:
+        block.sort(order='score')
+        block = block[::-1]
+
         if n_matches:
-            yield tuple(matches)[:n_matches]
+            yield block[:n_matches]
         else:
-            yield tuple(matches)
+            yield block