+++ dedupe/core.py
@@ -6,13 +6,11 @@ import sys
 if sys.version < '3':
     text_type = unicode
     binary_type = str
-    shelve_key = lambda x: x.encode()
     int_type = long
 else:
     text_type = str
     binary_type = bytes
     unicode = str
-    shelve_key = lambda x: x
     int_type = int
 
 import itertools
@@ -23,19 +21,15 @@ import operator
 import random
 import collections
 import warnings
-import shutil
-import shelve
-import pickle
+import functools
 
-
-try:
-    import collections.abc as collections_abc
-except ImportError:
-    import collections as collections_abc
 
 import numpy
 
 class ChildProcessError(Exception) :
+    pass
+
+class BlockingError(Exception):
     pass
 
 def randomPairs(n_records, sample_size):
@@ -101,7 +95,7 @@ def randomPairsWithReplacement(n_records, sample_size) :
     return [(p.item(), q.item()) for p, q in random_indices]
 
 
-class ScoreRecords(object) :
+class ScoreDupes(object) :
     def __init__(self, data_model, classifier, threshold) :
         self.data_model = data_model
         self.classifier = classifier
@@ -210,16 +204,16 @@ def scoreDuplicates(records, data_model, classifier, num_cores=1, threshold=0) :
 
     first, records = peek(records)
     if first is None:
-        raise ValueError("No records have been blocked together. "
-                         "Is the data you are trying to match like "
-                         "the data you trained on?")
+        raise BlockingError("No records have been blocked together. "
+                            "Is the data you are trying to match like "
+                            "the data you trained on?")
 
     record_pairs_queue = Queue(2)
     score_queue =  SimpleQueue()
     result_queue = SimpleQueue()
 
     n_map_processes = max(num_cores, 1)
-    score_records = ScoreRecords(data_model, classifier, threshold) 
+    score_records = ScoreDupes(data_model, classifier, threshold)
     map_processes = [Process(target=score_records,
                              args=(record_pairs_queue,
                                    score_queue))
@@ -241,7 +235,6 @@ def scoreDuplicates(records, data_model, classifier, num_cores=1, threshold=0) :
     if result :
         scored_pairs_file, dtype, size = result
         scored_pairs = numpy.memmap(scored_pairs_file,
-                                    mode='r',
                                     dtype=dtype,
                                     shape=(size,))
     else:
@@ -251,7 +244,6 @@ def scoreDuplicates(records, data_model, classifier, num_cores=1, threshold=0) :
     [process.join() for process in map_processes]
 
     return scored_pairs
-
 
 def fillQueue(queue, iterable, stop_signals) :
     iterable = iter(iterable)
@@ -299,6 +291,61 @@ def fillQueue(queue, iterable, stop_signals) :
             [queue.put(None) for _ in range(stop_signals)]
             break
 
+class ScoreGazette(object) :
+    def __init__(self, data_model, classifier, threshold) :
+        self.data_model = data_model
+        self.classifier = classifier
+        self.threshold = threshold
+
+    def __call__(self, block):
+        ids = []
+        records = []
+
+        for record_pair in block:
+            ((id_1, record_1, _), 
+             (id_2, record_2, _)) = record_pair
+
+            ids.append((id_1, id_2))
+            records.append((record_1, record_2))
+
+        distances = self.data_model.distances(records)
+        scores = self.classifier.predict_proba(distances)[:,-1]
+
+        mask = scores > self.threshold
+        id_type = sniff_id_type(ids)
+        ids = numpy.array(ids, dtype=id_type)
+
+        dtype = numpy.dtype([('pairs', id_type, 2),
+                             ('score', 'f4', 1)])
+
+        scored_pairs = numpy.empty(shape=numpy.count_nonzero(mask),
+                                   dtype=dtype)
+
+        scored_pairs['pairs'] = ids[mask]
+        scored_pairs['score'] = scores[mask]
+
+        return scored_pairs
+
+
+def scoreGazette(records, data_model, classifier, num_cores=1, threshold=0) :
+    if num_cores < 2 :
+        imap = map
+    else :
+        from .backport import Pool
+        n_map_processes = max(num_cores, 1)
+        pool = Pool(processes=n_map_processes)
+        imap = functools.partial(pool.imap_unordered, chunksize=1)
+
+    first, records = peek(records)
+    if first is None:
+        raise ValueError("No records to match")
+
+    score_records = ScoreGazette(data_model, classifier, threshold)
+
+    for scored_pairs in imap(score_records, records):
+        yield scored_pairs
+
+
 def peek(records) :
     try :
         record = next(records)
@@ -342,39 +389,6 @@ def Enumerator(start=0, initial=()):
         return collections.defaultdict(itertools.count(start).next, initial)
     except AttributeError : # py 3
         return collections.defaultdict(itertools.count(start).__next__, initial)
-
-
-class TempShelve(collections_abc.MutableMapping):
-    def __init__(self, filename):
-        self.path = tempfile.mkdtemp()
-        self.shelve = shelve.open(self.path + filename, 'n',
-                                  protocol=pickle.HIGHEST_PROTOCOL)
-
-    def close(self):
-        self.shelve.close()
-        shutil.rmtree(self.path)
-
-    def __getitem__(self, key):
-        key = shelve_key(key)
-        return self.shelve[key]
-
-    def __setitem__(self, key, value):
-        self.shelve[shelve_key(key)] = value
-
-    def __delitem__(self, key):
-        del self.shelve[shelve_key(key)]
-
-    def __iter__(self):
-        return iter(self.shelve)
-
-    def __len__(self):
-        return len(self.shelve)
-
-    def __contains__(self, key):
-        return shelve_key(key) in self.shelve
-
-    def values(self):
-        return viewvalues(self.shelve)
 
 
 def sniff_id_type(ids):